Use int instead of float for temp records
This should hopefully reduce the number of nuisance trips
on the low/high temps

+++ lcd_temperature_display.h
@@ -6,6 +6,7 @@
 #define FALSE 0
 
 //#define DEBUG
+//#define DEBUG_GETTEMPERATURE
 
 // Define constants
 #define AVERAGE_BINS_BITS 8   // 10 = 2^10 = 1024
@@ -33,7 +34,7 @@ struct ColorValue
 
 struct TemperatureRecord
 {
-  float temperature;
+  int temperature;
   unsigned long time;
 };
